Move _convert_responsive
As we now use it in layouts was causing a circular import
