+++ bokeh/plotting/figure.py
@@ -6,7 +6,8 @@ logger = logging.getLogger(__name__)
 from ..io import curdoc, curstate
 from ..models import Plot
 from ..models import glyphs, markers
-from .helpers import _get_range, _process_axis_and_grid, _process_tools_arg, _glyph_function, _convert_responsive
+from .helpers import _get_range, _process_axis_and_grid, _process_tools_arg, _glyph_function
+from ..util._plot_arg_helpers import _convert_responsive
 
 DEFAULT_TOOLS = "pan,wheel_zoom,box_zoom,save,reset,help"
