+++ bokeh/util/_plot_arg_helpers.py
@@ -1,0 +1,6 @@
+def _convert_responsive(responsive):
+    if responsive is True:
+        return 'scale_width'
+    if responsive is False:
+        return'fixed'
+    raise ValueError("'responsive' may only be True or False, passed %r" % responsive)
